[refactor] Calculator에서 Validation 분리
Calculator에 있던 Validation 기능들을 따로 분리 (관심사 분리)
